Move `montyformat` into repo Bench: 1124606
