+++ Cargo.toml
@@ -1,9 +1,11 @@
 [package]
 name = "monty"
 version = "1.0.0"
-edition = "2021"
-authors = ["Jamie Whiting, Viren & The Monty Authors"]
-rust-version = "1.83"
+edition = { workspace = true }
+authors = { workspace = true }
+rust-version = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
 
 [profile.release]
 panic = 'abort'
@@ -20,11 +22,8 @@ tunable = []
 value = []
 policy = []
 
-[workspace]
-members = ["datagen"]
-resolver = "2"
-
 [dependencies]
+montyformat = { workspace = true }
 rand = { version = "0.9.2", optional = true }
 rand_distr = { version = "0.5.1", optional = true }
 memmap2 = "0.9.5"
@@ -36,3 +35,19 @@ sha2 = "0.10.8"
 sha2 = "0.10.8"
 chrono = "0.4.38"
 zstd = { version = "0.13.2", features = ["zstdmt"] }
+
+[workspace]
+members = ["crates/datagen", "crates/montyformat"]
+resolver = "2"
+
+[workspace.package]
+authors = ["Jamie Whiting, Viren & The Monty Authors"]
+license = "AGPL-3.0"
+readme = "README.md"
+repository = "https://github.com/official-monty/Monty"
+rust-version = "1.83"
+edition = "2021"
+
+[workspace.dependencies]
+montyformat = { version = "0.10.0", path = "crates/montyformat" }
+monty = { path = "./" }
