+++ crates/datagen/src/rng.rs
