+++ crates/montyformat/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "montyformat"
+version = "0.10.0"
+description = "Chess implementation & compressed data formats for Monty"
+edition = { workspace = true }
+authors = { workspace = true }
+rust-version = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
